Sécurisation des routes et gestion des Rôles

+++ src/Controller/NoteController.php
@@ -10,6 +10,7 @@ use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Attribute\Route;
+use Symfony\Component\Security\Http\Attribute\IsGranted;
 
 #[Route('/note')]
 class NoteController extends AbstractController
@@ -17,14 +18,21 @@ class NoteController extends AbstractController
     #[Route('/', name: 'app_note_index', methods: ['GET'])]
     public function index(NoteRepository $noteRepository): Response
     {
+        if (!$this->getUser()) {
+            return $this->redirectToRoute('app_login');
+        }
         return $this->render('note/index.html.twig', [
             'notes' => $noteRepository->findAll(),
         ]);
     }
 
     #[Route('/new', name: 'app_note_new', methods: ['GET', 'POST'])]
+    #[IsGranted("ROLE_ADMIN", message: "Seul un admin peut ajouter une note")]
     public function new(Request $request, EntityManagerInterface $entityManager): Response
     {
+        if (!$this->getUser()) {
+            return $this->redirectToRoute('app_login');
+        }
         $note = new Note();
         $form = $this->createForm(NoteType::class, $note);
         $form->handleRequest($request);
@@ -45,14 +53,21 @@ class NoteController extends AbstractController
     #[Route('/{id}', name: 'app_note_show', methods: ['GET'])]
     public function show(Note $note): Response
     {
+        if (!$this->getUser()) {
+            return $this->redirectToRoute('app_login');
+        }
         return $this->render('note/show.html.twig', [
             'note' => $note,
         ]);
     }
 
     #[Route('/{id}/edit', name: 'app_note_edit', methods: ['GET', 'POST'])]
+    #[IsGranted("ROLE_ADMIN", message: "Seul un admin peut modifier une note")]
     public function edit(Request $request, Note $note, EntityManagerInterface $entityManager): Response
     {
+        if (!$this->getUser()) {
+            return $this->redirectToRoute('app_login');
+        }
         $form = $this->createForm(NoteType::class, $note);
         $form->handleRequest($request);
 
@@ -69,8 +84,12 @@ class NoteController extends AbstractController
     }
 
     #[Route('/{id}', name: 'app_note_delete', methods: ['POST'])]
+    #[IsGranted("ROLE_ADMIN", message: "Seul un admin peut supprimer une note")]
     public function delete(Request $request, Note $note, EntityManagerInterface $entityManager): Response
     {
+        if (!$this->getUser()) {
+            return $this->redirectToRoute('app_login');
+        }
         if ($this->isCsrfTokenValid('delete'.$note->getId(), $request->getPayload()->get('_token'))) {
             $entityManager->remove($note);
             $entityManager->flush();
